+++ lib/constants/string_constants.dart
@@ -4,7 +4,8 @@ class StringConstants {
       'https://explorer.walletconnect.com/?type=wallet';
 
   // TEXT
+  static const String connectButtonError = 'Network Error';
   static const String connectButtonIdle = 'Connect Wallet';
   static const String connectButtonConnecting = 'Connecting...';
-  static const String connectButtonAccount = 'Disconnect';
+  static const String connectButtonConnected = 'Disconnect';
 }