+++ test/walletconnect_modal_connect/walletconnect_modal_connect_ui_test.dart
@@ -1,19 +1,44 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
+import 'package:walletconnect_flutter_v2/walletconnect_flutter_v2.dart';
 import 'package:walletconnect_modal_flutter/constants/string_constants.dart';
 import 'package:walletconnect_modal_flutter/walletconnect_modal_flutter.dart';
 
 import '../mock_classes.mocks.dart';
 
+class WalletConnectModalServiceSpy extends MockWalletConnectModalService {
+  final List<VoidCallback> _listeners = [];
+
+  @override
+  void addListener(VoidCallback? listener) {
+    if (listener != null) {
+      _listeners.add(listener);
+    }
+  }
+
+  @override
+  void removeListener(VoidCallback? listener) {
+    _listeners.remove(listener);
+  }
+
+  @override
+  void notifyListeners() {
+    for (var listener in _listeners) {
+      listener();
+    }
+  }
+}
+
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
 
   group('WalletConnectModalService', () {
-    late MockWalletConnectModalService service;
+    late WalletConnectModalServiceSpy service;
 
     setUp(() async {
-      service = MockWalletConnectModalService();
+      service = WalletConnectModalServiceSpy();
+      when(service.initError).thenReturn(null);
       when(service.isConnected).thenReturn(false);
       when(service.isOpen).thenReturn(false);
     });
@@ -23,6 +48,7 @@ void main() {
       // FlutterError.onError = ignoreOverflowErrors;
 
       final GlobalKey key = GlobalKey();
+      // late BuildContext context;
 
       // Build our app and trigger a frame.
       await tester.pumpWidget(
@@ -31,9 +57,13 @@ void main() {
             body: SizedBox(
               width: 300,
               height: 100,
-              child: WalletConnectModalConnect(
-                key: key,
-                walletConnectModalService: service,
+              child: Builder(
+                builder: (context) {
+                  return WalletConnectModalConnect(
+                    key: key,
+                    service: service,
+                  );
+                },
               ),
             ),
           ),
@@ -42,7 +72,7 @@ void main() {
 
       await tester.pumpAndSettle();
 
-      // Connect flow
+      // idle state
       expect(
         find.text(
           StringConstants.connectButtonIdle,
@@ -51,26 +81,26 @@ void main() {
         findsOneWidget,
       );
       await tester.tap(find.byKey(key));
-
       await tester.pump();
 
       verify(
         service.open(context: anyNamed('context')),
       ).called(1);
 
+      // Connecting state
       when(service.isConnected).thenReturn(false);
       when(service.isOpen).thenReturn(true);
       service.notifyListeners();
 
-      await tester.pumpAndSettle();
+      await tester.pump();
 
-      // expect(
-      //   find.text(
-      //     StringConstants.connectButtonConnecting,
-      //     skipOffstage: false,
-      //   ),
-      //   findsOneWidget,
-      // );
+      expect(
+        find.text(
+          StringConstants.connectButtonConnecting,
+          skipOffstage: false,
+        ),
+        findsOneWidget,
+      );
 
       when(service.isConnected).thenReturn(true);
       when(service.isOpen).thenReturn(true);
@@ -78,10 +108,10 @@ void main() {
 
       await tester.pumpAndSettle();
 
-      // expect(
-      //   find.text(StringConstants.connectButtonAccount),
-      //   findsOneWidget,
-      // );
+      expect(
+        find.text(StringConstants.connectButtonConnected),
+        findsOneWidget,
+      );
 
       // Disconnect flow
       await tester.tap(find.byKey(key));
@@ -89,6 +119,54 @@ void main() {
       verify(
         service.disconnect(),
       ).called(1);
+    });
+
+    testWidgets('shows network error', (WidgetTester tester) async {
+      // FlutterError.onError = ignoreOverflowErrors;
+
+      when(service.initError).thenReturn(
+        const WalletConnectError(code: -1, message: 'Network error'),
+      );
+
+      final GlobalKey key = GlobalKey();
+      // late BuildContext context;
+
+      // Build our app and trigger a frame.
+      await tester.pumpWidget(
+        MaterialApp(
+          home: Scaffold(
+            body: SizedBox(
+              width: 300,
+              height: 100,
+              child: Builder(
+                builder: (context) {
+                  return WalletConnectModalConnect(
+                    key: key,
+                    service: service,
+                  );
+                },
+              ),
+            ),
+          ),
+        ),
+      );
+
+      await tester.pumpAndSettle();
+
+      // Got the init error
+      verify(service.initError).called(1);
+
+      expect(
+        find.text(
+          StringConstants.connectButtonError,
+          skipOffstage: false,
+        ),
+        findsOneWidget,
+      );
+
+      // Disabled button after tap
+      await tester.tap(find.byKey(key));
+      verifyNoMoreInteractions(service);
     });
   });
 }